fix: fireflies component
